Connect checkout connector
