+++ projects/core/src/cart/connectors/cart/cart.adapter.ts
@@ -1,5 +1,4 @@
 import { Observable } from 'rxjs';
-import { CheckoutDetails } from '../../../checkout/models/checkout.model';
 import { Cart } from '../../../model/cart.model';
 
 export abstract class CartAdapter {
@@ -23,17 +22,6 @@ export abstract class CartAdapter {
     cartId: string,
     details?: boolean
   ): Observable<Cart>;
-
-  /**
-   * Abstract method used to load checkout details
-   *
-   * @param userId
-   * @param cartId
-   */
-  abstract loadCheckoutDetails(
-    userId: string,
-    cartId: string
-  ): Observable<CheckoutDetails>;
 
   /**
    * Abstract method used to create cart. If toMergeCartGuid is specified, cart will be merged with existing one
