+++ projects/core/src/cart/connectors/cart/cart.connector.ts
@@ -2,7 +2,6 @@ import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs';
 
 import { CartAdapter } from './cart.adapter';
-import { CheckoutDetails } from '../../../checkout/models/checkout.model';
 import { Cart } from '../../../model/cart.model';
 
 @Injectable({
@@ -21,13 +20,6 @@ export class CartConnector {
     details?: boolean
   ): Observable<Cart> {
     return this.adapter.load(userId, cartId, details);
-  }
-
-  public loadCheckoutDetails(
-    userId: string,
-    cartId: string
-  ): Observable<CheckoutDetails> {
-    return this.adapter.loadCheckoutDetails(userId, cartId);
   }
 
   public create(
