+++ projects/core/src/checkout/connectors/checkout/checkout.connector.ts
@@ -2,6 +2,7 @@ import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs';
 import { CheckoutAdapter } from './checkout.adapter';
 import { Order } from '../../../model/order.model';
+import { CheckoutDetails } from '../../models/checkout.model';
 
 @Injectable({
   providedIn: 'root',
@@ -11,5 +12,12 @@ export class CheckoutConnector {
 
   public placeOrder(userId: string, cartId: string): Observable<Order> {
     return this.adapter.placeOrder(userId, cartId);
+  }
+
+  public loadCheckoutDetails(
+    userId: string,
+    cartId: string
+  ): Observable<CheckoutDetails> {
+    return this.adapter.loadCheckoutDetails(userId, cartId);
   }
 }
