+++ projects/core/src/occ/adapters/cart/occ-cart.adapter.spec.ts
@@ -9,7 +9,6 @@ import { ProductImageNormalizer } from '../../../occ/adapters/product/converters
 import { OccCartAdapter } from './occ-cart.adapter';
 import { ConverterService } from '../../../util/converter.service';
 import { CART_NORMALIZER } from '@spartacus/core';
-import { CheckoutDetails } from '../../../checkout/models/checkout.model';
 import { Occ } from '../../occ-models/occ.models';
 import { Cart } from '../../../model/cart.model';
 
@@ -27,12 +26,6 @@ const mergedCart: Cart = {
   name: 'mergedCart',
 };
 
-const checkoutData: CheckoutDetails = {
-  deliveryAddress: {
-    firstName: 'Janusz',
-  },
-};
-
 const usersEndpoint = '/users';
 const cartsEndpoint = 'carts';
 const BASIC_PARAMS =
@@ -45,8 +38,6 @@ const DETAILS_PARAMS =
   'totalPrice(formattedValue),totalItems,totalPriceWithTax(formattedValue),totalDiscounts(formattedValue),subTotal(formattedValue),' +
   'deliveryItemsQuantity,deliveryCost(formattedValue),totalTax(formattedValue),pickupItemsQuantity,net,' +
   'appliedVouchers,productDiscounts(formattedValue)';
-
-const CHECKOUT_PARAMS = 'deliveryAddress(FULL),deliveryMode,paymentInfo(FULL)';
 
 const MockOccModuleConfig: OccConfig = {
   backend: {
@@ -186,26 +177,6 @@ describe('OccCartAdapter', () => {
       );
       mockReq.flush({ carts: [cartData] });
       expect(result).toEqual(cartData);
-    });
-  });
-
-  describe('load checkout details', () => {
-    it('should load checkout details data for given userId, cartId', () => {
-      service.loadCheckoutDetails(userId, cartId).subscribe(result => {
-        expect(result).toEqual(checkoutData);
-      });
-
-      const mockReq = httpMock.expectOne(req => {
-        return (
-          req.method === 'GET' &&
-          req.url === `${usersEndpoint}/${userId}/${cartsEndpoint}/${cartId}`
-        );
-      });
-
-      expect(mockReq.cancelled).toBeFalsy();
-      expect(mockReq.request.responseType).toEqual('json');
-      expect(mockReq.request.params.get('fields')).toEqual(CHECKOUT_PARAMS);
-      mockReq.flush(checkoutData);
     });
   });
