+++ projects/core/src/occ/adapters/checkout/occ-checkout.adapter.ts
@@ -8,9 +8,11 @@ import { Occ } from '../../occ-models/occ.models';
 import { ConverterService } from '../../../util/converter.service';
 import { CheckoutAdapter } from '../../../checkout/connectors/checkout/checkout.adapter';
 import { ORDER_NORMALIZER } from '../../../checkout/connectors/checkout/converters';
+import { CheckoutDetails } from '@spartacus/core';
 
 // To be changed to a more optimised params after ticket: C3PO-1076
 const FULL_PARAMS = 'fields=FULL';
+const CHECKOUT_PARAMS = 'deliveryAddress(FULL),deliveryMode,paymentInfo(FULL)';
 
 @Injectable()
 export class OccCheckoutAdapter implements CheckoutAdapter {
@@ -23,6 +25,11 @@ export class OccCheckoutAdapter implements CheckoutAdapter {
   protected getOrderEndpoint(userId: string): string {
     const orderEndpoint = 'users/' + userId + '/orders';
     return this.occEndpoints.getEndpoint(orderEndpoint);
+  }
+
+  protected getCartEndpoint(userId: string): string {
+    const cartEndpoint = `users/${userId}/carts/`;
+    return this.occEndpoints.getEndpoint(cartEndpoint);
   }
 
   public placeOrder(userId: string, cartId: string): Observable<Order> {
@@ -39,5 +46,18 @@ export class OccCheckoutAdapter implements CheckoutAdapter {
       catchError((error: any) => throwError(error.json())),
       this.converter.pipeable(ORDER_NORMALIZER)
     );
+  }
+
+  loadCheckoutDetails(
+    userId: string,
+    cartId: string
+  ): Observable<CheckoutDetails> {
+    const url = this.getCartEndpoint(userId) + cartId;
+    const params = new HttpParams({
+      fromString: `fields=${CHECKOUT_PARAMS}`,
+    });
+    return this.http
+      .get<CheckoutDetails>(url, { params })
+      .pipe(catchError((error: any) => throwError(error)));
   }
 }
